Changed the name of the intercept back to `(Intercept)`. This matches outputs from lm.
Minor fixes for calculating the ridge penalty for lm models.

Re-ordered the passing of objects in createFTM. glm inherits lm, so was being incorrectly passed to ftmlm.

Changed the creation of FTM from glmnet objects. It now requires you to pass in the x and y matrix, rather than try to pull it from the environment.

Fixed a few minor bugs in subsetting, which was looking in the Xty/XtWz slots, rather than XtX/XtWX.

Created unit tests for main functions.